+++ Main.java
@@ -4,5 +4,6 @@ public abstract class Main {
 
     public static void main(String[] args) {
         System.out.println("Hello, World");
+        System.out.println(34 + 35);
     }
 }
