--- main.py
@@ -1,172 +1,0 @@
-#!/usr/bin/env python3
-
-"JVM Byte-code Parser in Python"
-
-import pprint
-import sys
-import io
-from typing import Any
-
-pp = pprint.PrettyPrinter()
-
-CONSTANT_Class              = 7
-CONSTANT_Fieldref           = 9
-CONSTANT_Methodref          = 10
-CONSTANT_InterfaceMethodref = 11
-CONSTANT_String             = 8
-CONSTANT_Integer            = 3
-CONSTANT_Float              = 4
-CONSTANT_Long               = 5
-CONSTANT_Double             = 6
-CONSTANT_NameAndType        = 12
-CONSTANT_Utf8               = 1
-CONSTANT_MethodHandle       = 15
-CONSTANT_MethodType         = 16
-CONSTANT_InvokeDynamic      = 18
-
-class_access_flags = [
-    ("ACC_PUBLIC"     , 0x0001),
-    ("ACC_FINAL"      , 0x0010),
-    ("ACC_SUPER"      , 0x0020),
-    ("ACC_INTERFACE"  , 0x0200),
-    ("ACC_ABSTRACT"   , 0x0400),
-    ("ACC_SYNTHETIC"  , 0x1000),
-    ("ACC_ANNOTATION" , 0x2000),
-    ("ACC_ENUM"       , 0x4000)
-]
-
-method_access_flags = [
-    ("ACC_PUBLIC"       , 0x0001),
-    ("ACC_PRIVATE"      , 0x0002),
-    ("ACC_PROTECTED"    , 0x0004),
-    ("ACC_STATIC"       , 0x0008),
-    ("ACC_FINAL"        , 0x0010),
-    ("ACC_SYNCHRONIZED" , 0x0020),
-    ("ACC_BRIDGE"       , 0x0040),
-    ("ACC_VARARGS"      , 0x0080),
-    ("ACC_NATIVE"       , 0x0100),
-    ("ACC_ABSTRACT"     , 0x0400),
-    ("ACC_STRICT"       , 0x0800),
-    ("ACC_SYNTHETIC"    , 0x1000),
-]
-
-def parse_flags(value: int, flags: list[tuple[str, int]]) -> list[str]:
-    "Parse flags from value and return names of flags set"
-    return [name for (name, mask) in flags if (value&mask) != 0]
-
-def parse_u1(f: io.BufferedReader) -> int: return int.from_bytes(f.read(1), 'big')
-def parse_u2(f: io.BufferedReader) -> int: return int.from_bytes(f.read(2), 'big')
-def parse_u4(f: io.BufferedReader) -> int: return int.from_bytes(f.read(4), 'big')
-
-def parse_attributes(f: io.BufferedReader, count: int) -> list[dict[str, int | bytes]]:
-    "Parse attributes from f"
-    attributes = []
-    for _ in range(count):
-        # attribute_info {
-        #     u2 attribute_name_index;
-        #     u4 attribute_length;
-        #     u1 info[attribute_length];
-        # }
-        attribute: dict[str, int | bytes] = {}
-        attribute['attribute_name_index'] = parse_u2(f)
-        attribute_length = parse_u4(f)
-        attribute['info'] = f.read(attribute_length)
-        attributes.append(attribute)
-    return attributes
-
-def parse_class_file(file_path: str) -> dict[str, Any]:
-    "Parse class f given it's path"
-    with open(file_path, "rb") as f:
-        clazz: dict[str, Any] = {}
-        clazz['magic'] = hex(parse_u4(f))
-        clazz['minor'] = parse_u2(f)
-        clazz['major'] = parse_u2(f)
-        constant_pool_count = parse_u2(f)
-        constant_pool = []
-        for _ in range(constant_pool_count-1):
-            cp_info: dict[str, Any] = {}
-            tag = parse_u1(f)
-            if tag == CONSTANT_Methodref:
-                cp_info['tag'] = 'CONSTANT_Methodref'
-                cp_info['class_index'] = parse_u2(f)
-                cp_info['name_and_type_index'] = parse_u2(f)
-            elif tag == CONSTANT_Class:
-                cp_info['tag'] = 'CONSTANT_Class'
-                cp_info['name_index'] = parse_u2(f)
-            elif tag == CONSTANT_NameAndType:
-                cp_info['tag'] = 'CONSTANT_NameAndType'
-                cp_info['name_index'] = parse_u2(f)
-                cp_info['descriptor_index'] = parse_u2(f)
-            elif tag == CONSTANT_Utf8:
-                cp_info['tag'] = 'CONSTANT_Utf8'
-                length = parse_u2(f)
-                cp_info['bytes'] = f.read(length)
-            elif tag == CONSTANT_Fieldref:
-                cp_info['tag'] = 'CONSTANT_Fieldref'
-                cp_info['class_index'] = parse_u2(f)
-                cp_info['name_and_type_index'] = parse_u2(f)
-            elif tag == CONSTANT_String:
-                cp_info['tag'] = 'CONSTANT_String'
-                cp_info['string_index'] = parse_u2(f)
-            else:
-                assert False, f"Unexpected tag {tag}"
-            constant_pool.append(cp_info)
-        clazz['constant_pool'] = constant_pool
-        clazz['access_flags'] = parse_flags(parse_u2(f), class_access_flags)
-        clazz['this_class'] = parse_u2(f)
-        clazz['super_class'] = parse_u2(f)
-        interfaces_count = parse_u2(f)
-        interfaces = []
-        for _ in range(interfaces_count):
-            assert False, "Parsing interfaces is not implemented"
-        clazz['interfaces'] = interfaces
-        fields_count = parse_u2(f)
-        fields = []
-        for _ in range(fields_count):
-            assert False, 'Parsing fields is not implemented'
-        clazz['fields'] = fields
-        methods_count = parse_u2(f)
-        methods = []
-        for _ in range(methods_count):
-            # u2             access_flags;
-            # u2             name_index;
-            # u2             descriptor_index;
-            # u2             attributes_count;
-            # attribute_info attributes[attributes_count];
-            method: dict[str, Any] = {}
-            method['access_flags'] = parse_flags(parse_u2(f), method_access_flags)
-            method['name_index'] = parse_u2(f)
-            method['descriptor_index'] = parse_u2(f)
-            attributes_count = parse_u2(f)
-            method['attributes'] = parse_attributes(f, attributes_count)
-            methods.append(method)
-        clazz['methods'] = methods
-        attributes_count = parse_u2(f)
-        clazz['attributes'] = parse_attributes(f, attributes_count)
-        return clazz
-
-def find_methods_by_name(clazz: dict[str, Any], name: bytes) -> list[dict[str, Any]]:
-    "Return methods from parsed class by name"
-    return [method
-            for method in clazz['methods']
-            if clazz['constant_pool'][method['name_index'] - 1]['bytes'] == name]
-
-def find_attributes_by_name(clazz: dict[str, Any],
-                            attributes: list[dict[str, int]],
-                            name: bytes) ->  list[dict[str, int]]:
-    "Return attributes from parsed class by name"
-    return [attr
-            for attr in attributes
-            if clazz['constant_pool'][attr['attribute_name_index'] - 1]['bytes'] == name]
-
-def cli_run() -> None:
-    "Run from command line"
-    filename = './Main.class'
-    arguments = sys.argv[1:]
-    if arguments:
-        filename = arguments[0]
-    clazz = parse_class_file(filename)
-    pp.pprint(clazz)
-
-if __name__ == '__main__':# Do not run if imported as module
-    cli_run()
